requirements.txt
fastapi
uvicorn
boto3
chromadb
langchain
langchain-community
langchain-google-genai
huggingface-hub
sentence-transformers
PyMuPDF
python-multipart
python-dotenv
requests
httpx
orjson
pydantic
pydantic-settings
scikit-learn
numpy
torch
transformers
pytest
moto
localstack-client
awscli